refactor(Modules): moves the modules around into a more logical order to make contribution easier
