+++ src/output/fmt.rs
