+++ src/output/mod.rs
@@ -1,0 +1,7 @@
+mod help;
+mod usage;
+
+pub mod fmt;
+
+pub use self::help::Help;
+pub use self::usage::Usage;
