+++ src/parse/matches/arg_matches.rs
@@ -9,8 +9,7 @@ use indexmap::IndexMap;
 
 // Internal
 use INVALID_UTF8;
-use args::MatchedArg;
-use args::SubCommand;
+use parse::{MatchedArg, SubCommand};
 
 /// Used to get information about the arguments that where supplied to the program at runtime by
 /// the user. New instances of this struct are obtained by using the [`App::get_matches`] family of
