+++ src/parse/mod.rs
@@ -1,0 +1,13 @@
+pub mod errors;
+pub mod features;
+
+mod matches;
+mod arg_matcher;
+mod parser;
+mod validator;
+
+pub use self::parser::{Parser, ParseResult};
+pub use self::matches::ArgMatches;
+pub use self::arg_matcher::ArgMatcher;
+pub use self::matches::{Values, OsValues, SubCommand, MatchedArg};
+pub use self::validator::Validator;
