+++ src/util/mod.rs
@@ -1,0 +1,9 @@
+mod map;
+mod osstringext;
+mod strext;
+
+pub use self::map::{Values, VecMap};
+pub use self::osstringext::OsStrExt2;
+#[cfg(any(target_os = "windows", target_arch = "wasm32"))]
+pub use self::osstringext::OsStrExt3;
+pub use self::strext::_StrExt;
