+++ src/util/osstringext.rs
